% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cov_columns.R
\name{cov_columns}
\alias{cov_columns}
\title{Append columns related to the position of peptide in a protein: Start (start position of a peptide in a protein), End (end position of a peptide in a protein), Protlength (length of the full protein), Appearance (if a peptide appears multiple times in a protein).}
\usage{
cov_columns(df, elementid = "Peptide")
}
\arguments{
\item{elementid}{Colname of column containing peptide sequences. Will be searched against the sequences extracted from the proteome file.}

\item{data_table}{Data frame to which the coverage columns will be appended. Must contain a column of peptide sequences and a column of the parental protein of each peptide.}
}
\value{
Data frame with the following added columns: "Start", "End", "Sequence", "Appearance".
}
\description{
Append columns related to the position of peptide in a protein: Start (start position of a peptide in a protein), End (end position of a peptide in a protein), Protlength (length of the full protein), Appearance (if a peptide appears multiple times in a protein).
}
\examples{
cov_columns(test_data, seqid='ID',elementid='Peptide')
}
