% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fasta_tidier.R
\name{fasta_tidier}
\alias{fasta_tidier}
\title{Converts a SeqFastadna object into a tidy dataframe.}
\usage{
fasta_tidier(fasta_seqinr_object, annot = FALSE)
}
\arguments{
\item{fasta_seqinr_object}{FASTA file read by seqinr::read.fasta from which the sequences will be extracted.}

\item{annot}{If TRUE, extract sequence ID annotation into its own column}

\item{...}{Unused extra arguments}
}
\value{
A \code{tidy} data.frame with one column for ID and one column for sequence.
\item{ID}{Sequence header from FASTA '>ID' }
\item{Sequence}{Corresponding sequence, any bioalphabet (RNA/DNA/etc)}
\item{Annot}{Full annotated FASTA header '>ID info|info|info'}
}
\description{
Converts a SeqFastadna object into a tidy dataframe.
}
\examples{
seqinr::read.fasta function reads a FASTA file (format below) and returns a SeqFastadna object
   > GENENAME1
   atgctgacgta
   agctacagtat
   cagctactaag
seqinr_fasta_object <- seqinr::read.fasta("test.fasta")
fasta_tidier(seqinr_fasta_object)
fasta_tidier(seqinr_fasta_object, annot=TRUE)
}
