% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pep_elut_data.R
\docType{data}
\name{msdata}
\alias{msdata}
\title{Fractionation data}
\format{An object of class data.frame}
\usage{
data(msdata)
}
\description{
Data from a fractionation experiment
}
\examples{
data(msdata)
}
\keyword{datasets}
