% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalit.R
\name{normalit}
\alias{normalit}
\title{Normalizes a vector of values from 0 to 1}
\usage{
normalit(m)
}
\arguments{
\item{m}{A vector}

\item{...}{Unused extra arguments}
}
\value{
A normalized vector with values rescaled from 0 to 1
}
\description{
Normalizes a vector of values from 0 to 1
}
\examples{
normalit(m)
}
