% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_fasta.R
\name{read_fasta}
\alias{read_fasta}
\title{Read a FASTA file into a tidy dataframe. Wrapper around seqinr::read.fasta}
\usage{
read_fasta(fasta_filename, annot = FALSE)
}
\arguments{
\item{fasta_filename}{FASTA file from which the sequences will be extracted.}

\item{annot}{If TRUE, extract sequence ID annotation into its own column}

\item{...}{Unused extra arguments}
}
\value{
A \code{tidy} data.frame with one column for ID and one column for sequence.
\item{ID}{Sequence header from FASTA '>ID' }
\item{Sequence}{Corresponding sequence, any bioalphabet (RNA/DNA/etc)}
\item{Annot}{Full annotated FASTA header '>ID info|info|info'}
}
\description{
Read a FASTA file into a tidy dataframe. Wrapper around seqinr::read.fasta
}
\examples{
test.fasta is a FASTA formatted file for ex:
   > GENENAME1
   atgctgacgta
   agctacagtat
   cagctactaag
read_fasta("test.fasta")
read_fasta("test.fasta", annot=TRUE)
}
