% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/complete_counts.R
\name{complete_counts}
\alias{complete_counts}
\title{Fill missing x-axis values with zero.}
\usage{
complete_counts(raw_data, xaxis, yaxis)
}
\arguments{
\item{raw_data}{Data frame}

\item{xaxis}{Name of the column with X-axis values}

\item{yaxis}{Name of the column with Y-axis values}

\item{outname}{If export = TRUE, output the resulting .}
}
\value{
Data frame with added rows
}
\description{
Fill missing x-axis values with zero.
}
\examples{
complete_counts(test_data,xaxis='FractionID',yaxis='PeptideCount')
}
