% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cov_columns.R
\name{cov_columns}
\alias{cov_columns}
\title{Append columns related to the coverage of a protein.}
\usage{
cov_columns(data_table, proteome, groupid, elementid,
  outname = paste0(deparse(substitute(data_table)), "_cov.csv"),
  export = FALSE)
}
\arguments{
\item{data_table}{Data frame to which the coverage columns will be appended.}

\item{proteome}{FASTA file from which the sequences will be extracted.}

\item{groupid}{Data frame column (factor) corresponding to the protein name to be matched with the IDs extracted from the FASTA file.}

\item{elementid}{Data frame column (factor) corresponding to the peptide sequence to be searched against the sequences extracted from the FASTA file.}

\item{outname}{If export = TRUE, output the resulting data frame as a .csv file named after this argument.}

\item{export}{If false (default), return a data frame. If true, write a .csv file named after the input file + '_cov.csv'.}
}
\value{
Data frame with the following added columns: "Start", "End", "Sequence", "Appearance".
}
\description{
Append columns related to the coverage of a protein.
}
\examples{
test_data <- read.csv(paste0(system.file('extdata',package='pepliner'),'/msdata.csv'))
sequences <- paste0(system.file('extdata',package='pepliner'),'/proteome.fasta')
cov_columns(test_data,sequences,groupid='ID',elementid='Peptide')
}
