% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cov_columns.R
\name{cov_columns}
\alias{cov_columns}
\title{Append columns related to the position of peptide in a protein: Start (start position of a peptide in a protein), End (end position of a peptide in a protein), Protlength (length of the full protein), Appearance (if a peptide appears multiple times in a protein).}
\usage{
cov_columns(data_table, proteome, groupid, elementid)
}
\arguments{
\item{data_table}{Data frame to which the coverage columns will be appended. Must contain a column of peptide sequences and a column of the parental protein of each peptide.}

\item{proteome}{FASTA file from which the sequences will be extracted. IDs must match protein IDs in the input data_table.}

\item{groupid}{Colname of column (factor, not numeric) containing protein IDs. IDs must match proteome identifiers.}

\item{elementid}{Colname of column containing peptide sequences. Will be searched against the sequences extracted from the FASTA file.}
}
\value{
Data frame with the following added columns: "Start", "End", "Sequence", "Appearance".
}
\description{
Append columns related to the position of peptide in a protein: Start (start position of a peptide in a protein), End (end position of a peptide in a protein), Protlength (length of the full protein), Appearance (if a peptide appears multiple times in a protein).
}
\examples{
test_data <- read.csv(system.file('extdata/msdata.csv',package='pepliner'))
sequences <- system.file('extdata/proteome.fasta',package='pepliner')
cov_columns(test_data,sequences,groupid='ID',elementid='Peptide')
}
