% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covplot.R
\name{covplot}
\alias{covplot}
\title{Make a coverage plot for all the peptides retrieved for a single protein.}
\usage{
covplot(input_data, group_name = "", sort_column = "Start", groupid = "",
  elementid)
}
\arguments{
\item{input_data}{Data frame to which the coverage columns will be appended.}

\item{group_name}{(Optional) If specified, make plots from only those rows whose ID matched this argument.}

\item{sort_column}{(Optional) If specified, rows in the final plot will be sorted according to this (numeric) column.}

\item{groupid}{Data frame column (factor) corresponding to the names of the groups of elements in the data.}

\item{elementid}{Data frame column (factor) corresponding to the elements for each of which a row of the final plot will be produced.}
}
\value{
ggplot item to be plotted using cowplot::ggdraw().
}
\description{
Make a coverage plot for all the peptides retrieved for a single protein.
}
\examples{
'ms_data.csv' \%>\% complete_counts() \%>\% covplot(groupid='ID',elementid='Peptide',sort_column='Start') \%>\% ggdraw
}
