% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covplot.R
\name{covplot}
\alias{covplot}
\title{Make a coverage plot for all the peptides retrieved for a single protein.}
\usage{
covplot(input_data, elementid, groupid = "", group_name = "",
  sort_column = "Start")
}
\arguments{
\item{input_data}{Data frame to which the coverage columns will be appended.}

\item{elementid}{Data frame column (factor) corresponding to the elements for each of which a row of the final plot will be produced.}

\item{groupid}{Data frame column (factor) corresponding to the names of the groups of elements in the data.}

\item{group_name}{(Optional) If specified, make plots from only those rows whose ID matched this argument.}

\item{sort_column}{(Optional) If specified, rows in the final plot will be sorted according to this (numeric) column.}
}
\value{
ggplot item to be plotted using cowplot::ggdraw().
}
\description{
Make a coverage plot for all the peptides retrieved for a single protein.
}
\examples{
library(purrr)
test_data <- read.csv(paste0(system.file('extdata',package='pepliner'),'/msdata.csv'))
sequences <- paste0(system.file('extdata',package='pepliner'),'/proteome.fasta')
cov_columns(test_data,sequences,groupid='ID',elementid='Peptide') \%>\%
covplot('Peptide','ID','sp|P55036|PSMD4_HUMAN') \%>\%
cowplot::ggdraw()
}
