% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparkplot.R
\name{normalit}
\alias{normalit}
\title{Normalize values in a vector. Auxiliary function.}
\usage{
normalit(m)
}
\arguments{
\item{m}{A vector}
}
\description{
Each sparkline row should have values scaled from 0 to 1.
}
