% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_fasta.R
\name{read_fasta}
\alias{read_fasta}
\title{Read a fasta file into a two column dataframe, with one column for sequenceid and one column for sequence}
\usage{
read_fasta(proteome_file, seqid = "ID", seq = "Sequence")
}
\arguments{
\item{proteome_file}{FASTA file from which the sequences will be extracted.}

\item{seqid}{Colname of column for containint protein IDs.}

\item{seq}{Colname for column containing sequences. Will be searched against the sequences extracted from the FASTA file.}
}
\value{
Two column dataframe with one column for ID and one column for sequence.
}
\description{
Read a fasta file into a two column dataframe, with one column for sequenceid and one column for sequence
}
\examples{
read_fasta(test.fasta, id="ID", sequenceid = "Sequence")
}
