% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparkplot.R
\name{sparkplot}
\alias{sparkplot}
\title{Make a line plot for each element of a group.}
\usage{
sparkplot(input_data, elementid, xaxis, yaxis, groupid = "",
  group_name = "", condit = "", sort_column = "")
}
\arguments{
\item{input_data}{Data frame to which the coverage columns will be appended.}

\item{elementid}{Data frame column (factor) corresponding to the elements for each of which a row of the final plot will be produced.}

\item{xaxis}{X-axis values for each sub-plot. Must match across elements.}

\item{yaxis}{Y-axis values for each sub-plot. The final plot normalizes these values for each row.}

\item{groupid}{Data frame column (factor) corresponding to the names of the groups of elements in the data.}

\item{group_name}{(Optional) If specified, make plots from only those rows whose ID matched this argument.}

\item{condit}{(Optional) Name of the condition column from the data frame. Data points corresponding to different values of condit will be plotted with a different color/linetype.}

\item{sort_column}{(Optional) If specified, rows in the final plot will be sorted according to this (numeric) column.}
}
\value{
ggplot item to be plotted using cowplot::ggdraw().
}
\description{
Make a line plot for each element of a group.
}
\examples{
library(purrr)
test_data <- read.csv(paste0(system.file('extdata',package='pepliner'),'/msdata.csv'))
sequences <- paste0(system.file('extdata',package='pepliner'),'/proteome.fasta')
cov_columns(test_data,sequences,groupid='ID',elementid='Peptide') \%>\%
complete_counts('FractionID','PeptideCount') \%>\%
sparkplot('Peptide','FractionID','PeptideCount','ID',
    group_name='sp|P55036|PSMD4_HUMAN',sort_column='Start') \%>\%
cowplot::ggdraw()
}
